requirements.txt
customtkinter
mysql-connector-python
PyMuPDF
Faker